[Tool Calls] Match pending side panel to running tool UX
